Codespaces compatibility improvements
Made a "transparent" compatibility hack that allows you to use websockets in Codespaces.
It relies on rewriting the codespaces host url, which is not very good. But it works for now.

Moved the "websocket over http" setting to a environment variable in the server.

This entire change was developed inside Codespaces itself.
It only crashed a few times while doing this work.