Generated protocol, nested probes, and minor QOL
Moved many datatypes that previously had duplicate implementations
across client and server into  a shared protocol specification, that
is used to automatically generate both Java and TypeScript sources.

Added the possibility to nest probes. When the result of a probe
is a node, you can now click a tiny arrow to open an inline probe inside
the existing window. When changes happen, the inline probe is evaluated
on the result of the outer probe. This allows you to construct more
advanced probe chains than could be done before. Nested probes can have
nested probes as well.

Various quality of life changes, including:
* Ability to save all tests as expected with a single click, instead of
  having to do it individually.
* Improved test-diff detection, with the help of the npm package 'onp'.
* Added ability to save window state into URL. This lets you share state
  with others, or save multiple 'workspaces' for example as bookmarks.