Experiment support for "query|meta|set|coll"-probe
Name tbd.

Add "*.propName" in the property selector window, and a new kind of
probe will be created. This probe fins all nodes that exist beneath the
current point in the AST and that has a property called "propName". It
then creates an inline window for each such probe. Can be used to
mass-query for a specific property. Nodes can opt-out of showing up in
this by implementing "cpr_propSearchVisible" and returning false. That
property is also naming tbd.