"Oneshot" request support
Added support for running individual "/wsput" requests as command-line
arguments to codeprober. Format for running:

  java -jar code-prober.jar --oneshot='query_here' --output='file_here'

The query will be processed as if it was sent to the /wsput endpoint.
The result is written to 'file_here', and then CodeProber shuts down.
I.e the normal web/websocket server doesn't start.