Add authorization token support
When CodeProber starts, it will now print a url that includes a
authorization token. It will say something like:

 "visit 'http://localhost:8001?auth=key-MS4wNTc0 in your browser"

When connecting from non-localhost, that key will be used for
authorization purposes. It should improve the experience for people
running CodeProber in WSL, as you no longer have to set
PERMIT_REMOTE_CONNECTIONS=true.