Bidirectional backing_file syncing
The backing_file feature now syncs from the file system to the
CodeProber client, in addition to the client->fs syncing that existed
before.
You can now in practice use any editor you wish with CodeProber. Make
changes in that editor, save, and watch the probes in CodeProber get
updated.