Improvements to SessionLogger
Add a log event on startup that includes the version of CodeProber

Add a log event event when a client connects that includes the
protocol they connect with (http or ws).

Reduced the AUTO_STOP_DATA_THRESHOLD to 32mb. There is a risk of the
file being rejected by hosting providers if it is too big. For example
GitHub blocks files above 100mb, which is smaller than the previous
threshold. 32mb should be safer.

Made it possible to log to a persistent file across sessions, instead
of always generating a new session id and new files for each flush.
This is to reduce the need for "git add logs/*" after using CodeProber.