Fixed worker process concurrency issue
Fixed an issues that could lead to the message "Got non-json
data from worker" being printed. To explain how this could happen,
them we'll use the following terminology:

- Coordinator: the main CodeProber process.
- Worker: worker processes that communicate with the coordinator via
  stdout.

Workers only exist when setting "--concurrent=N" while running CodeProber.
Coordinator and worker processes comunicate via standard I/O streams, in
the form of JSON messages.
The core issue this commit solves is that the JSON got corrupted by
non-json data due to concurrency issues.

To prevent "random" logs from CodeProber itself or the underlying
tool from being interpreted as messages to the coordinator, we use some
intereception code that forwards the data to a file log (if an
environment variable is set), or simply discards it.

The interceptor implementation (StreamInterceptor) by default ignores
data written on "other" threads, to avoid intercepting unintended
messages. This is usuaally intended behavior. However, in this case it
meant that prints from the underlying tool (which is handled on a
background thread), could reach standard out and be interpreted as
messages to the coordinator, whether or not they were actual JSON
messages.

This was usually not an issue, as all messages to the coordinator have
to be formatted a certain way to be interpreted as actual messages, and
we print the formatting header and body all within a "synchronized(..)"
block, to prevent any intermingling of data. However, not all
PrintStream implementations seem to internally synchronize as well, as
has been the case for us during development.
For example, this synchronizes on "this" in write(int):
https://github.com/openjdk-mirror/jdk7u-jdk/blob/master/src/share/classes/java/io/PrintStream.java
..but this doesn't:
https://developer.classpath.org/doc/java/io/PrintStream-source.html

This commit changes the interception code to intercept data on *all*
threads in worker processes.
Also it adds some overdue test cases for StreamInterceptor.