Reset AST on errors and parse
For many calls to the underlying tool, CodeProber will wrap interactions
within a try/catch block, and present any potential error in the
CodeProber UI. For most cases, this is fine. However, some
errors may leave the AST in an invalid state, which causes future
evaluations be incorrect unless the state is explicitly reset.

This commit adds two levels of "resetting":
 when an exception is thrown
during property evaluation:

1) When an exception is thrown during property evaluation, the cached
   AST is discarded, which forces a fresh parse for the next request.

2) For every fresh parse, CodeProber calls
   "<TheRootNode>.state().reset()". In JastAdd, this resets a state
   related to circular attributes which otherwise wouldn't be reset by a
   fresh parse, because it is static.

Some guardrails are added for case 2 to avoid doing this on non-JastAdd
tools. If necessary, it can be disabled entirely by setting the system
property "cpr.resetASTStateOnParse" to anything except "true".

Added tests for the new functionality too.