Improved support for new use cases
Made several changes that improve support for using CodeProber in new
ways, or with new (non-JastAdd) languages.

It is now possible to host the CodeProber frontend in conjunction with a
plugin that assigns window.CPR_REQUEST_HANDLER, and then handles
requests locally. This can pass requests to any arbitrary backend, or it
can handle the requests inside the browser, making for a completely
client-side CodeProber experience.

Added improvements for supporting "labeled" properties in more places,
and more extension points for describing the AST structure without
relying on JastAdd annotations. For example:
- cpr_lGetChildName can be used to describe which properties are child
  accessors, and what their names are.
- cpr_lGetAspectName can be used to describe which aspect a given property
  belongs to.
- cpr_describeParentConnection can be used to manually provide steps
  during node locator construction, simplifying/enabling NTA location
  for non-JastAdd tools.

Added Protocol generation for a language called KL. This is a hobby
language under development which potentially will be tightly integrated
with CodeProber in the future. This protocol generation is the first
step.