Rename CodeProber.jar and release via Github
This commit makes two long overdue and slightly breaking changes:

1) Renames "code-prober.jar" and "code-prober-dev.jar" to "CodeProber.jar"
2) Changes release management to be based on normal github releases,
   instead of the previous "VERSION" file and jar in git history.

The tool is called CodeProber, not Code-Prober, so the old jar name was
simply incorrect. Also, separating release and dev builds is
unnecessary, as that metadata is already present inside the build itself
(in cpr.properties).

Switching to github releases should stop the size of the repository from
growing too quickly.