More lenient tracing, experimental HTML support
Improved TracingBuilder to not be so strict. It no longer calls
System.exit when it is unable to create locators. In addition, it can be
configured to skip certain AST types or attribute names.

Added experimental support for HTML output in probes. String values that
start with "@@HTML:" are presented in an iframe on the client side,
similar to how values starting with "@@DOT:" are presented as if they
are a dot graph.