Workspace and Text Probe support
This commit adds two major features to CodeProber: support for Workspace
and text probes.

== Workspace support

A "workspace" is a virtual filesystem visible within CodeProber.
It always contains one "temp file", which represents the in-browser temp
file that has existed since the beginning of CodeProber.
It may contain additional files, backed by a real filesystem.

Each entry in the workspace has its own list of active probes. When
jumping between files, the probes get saved/loaded accordingly.
The active probes for the non-temp file are also stored on the
filesystem, inside a ".cpr" directory next the real file.
For example, the real file system may look like this:

  workspace/
      foo
      bar
      .cpr/
          foo

In this case, the workspace has two files visible in CodeProber.
One of the files ("foo") additionally has some associated metadata
with it, likely due to the user having active probes in that file.

Users may want to add ".cpr" to their gitignore to keep their active
probes private. However, it you wish to share probe configurations
within a project, consider adding it to git too.

A workspace is specified with the system property "cpr.workspace".
You can optionally specify "cpr.workspaceFilePattern", which is a
pattern that filters which files should be visible in the CodeProber UI.
Example invocation with these set:

 java -Dcpr.workspace=ws \
      -Dcpr.workspaceFilePattern=".*\.java" \
      -jar codeprober.jar mytool.jar

== Text Probes

Text probes (name is work in progress) are "probes" that exist entirely
as text within the source file.
CodeProber detects these probes by finding the following pattern:

    [[NodeType[index].attrName1.attrName2.etc=expectedVal]]

- "NodeType" is the name of an AST node type.
- "index" is optional, used to disambiguate in case there are multiple
  nodes of the specified type on that line.
- "attrName1.attrName2.etc" is a chain of property names. All but the
  last one in the chain must resolve to an AST node reference, similar
  to normal nested probes inside CodeProber.
- "expectedVal" is a string representation of the expected result of the
  probe. This can be left out to "just" probe.

The CodeProber editor adds green, red or blue boxes around text probes
to signify "successful", "fail" and "info", respectively.
  - "success"/"fail" is used when an "expectedVal" is or is not met.
  - "info" is used when no "expectedVal" exists.

A comparison can optionally include "!" and "~", which mean "not" and
"contains" respectively.

Some example text probes and their possible meaning are listed below.
Their exact meaning depend on the specific compiler/analyzer implementation:

  "Type of the Num node should be int"
  [[Num.type=int]]

  "The function being called should expected 2 parameters"
  [[CallExpr.decl.numParameters=2]]

  "There should be a message in Program.errors which contains the string
   'duplicate declaration'"
  [[Program.errors~=duplicate declaration]]

  "Program.errors should not be empty"
  [[Program.errors!=[]]]

Text probes are possibly going to replace the "normal" test probes that were
prototyped 2 years ago (and still lives in the codebase).
Normal test probes may be removed, or significantly refactored in the
near future.

The "--test" cli argument to CodeProber behaves differently if a
workspace is activated. Instead of loading tests from "cpr.testDir"
as before, it will look for all text probes in the workspace,
evaluate their assertions and set an exit code appropriately
if there are any assertion failures.