Convert to/from text probe format
Hovering a text probe now displays an option "Open Normal Probe", which
opens a normal probe when clicked.

Normal probes now has a "Copy as text probe" option in the 3-dot menu.
Note, this option is only available for probes that are possible to
express in text probe form. This excludes nodes that are identified by
parameterized higher order attributes.