Node locator and tracing improvements
Perform a significant number of improvements to performance to node locator,
as well as improvements to the kind of arguments that are supported by node locators.
Also made improvements to the tracing feature.

Added custom binary serialization and deserialization of the protocol format.
This provides a more compact way to store protocol objects, and a way to compare them.

This commit should provide an overall significant performance boost when performing large
amounts of tracing and locator construction programmatically. You are not likely to notice any
significant improvement in normal, human-based interaction as it was already fast enough there.

For some use cases, there is a possibility to get even more performance by assigning
`CreateLocator.identityLocatorCache` to a non-null value. As the name implies, this is a node
locator construction cache. If you are using mutable ASTs (whether intentional, or due to
implementation errors), then this feature will cause incorrect locators to be created.
Therefore, this is set to null by default.