TextProbe variable support
Added support for "variables" in text probes. Variables may be used on
either side of a comparison.

Examples of new probes now supported and their meanings are:
  [[$x:=Y.z]] -> Define variable $x to be equal to the query "Y.z"
  [[$x=$a.b]] -> Assert that variable x resolves to b on variable $a.
  [[$x=Y.z]]  -> Assert that variable x resolves to the string "Y.z"
  [[Y.z=$x]]  -> Assert that "Y.z" resolves to the same as variable x

Note that the two last examples are mirrored, but do not have the same
meaning. In other words, '=' is not commutative. The left and right
side of a comparison are interpreted differently.
* The left side is always a "query", e.g. finding a node and optionally
  resolving a list of attributes. The node may be a variable, or the
  name of a type of variable on the current line.
* The right side is treated as a query IF it starts with "$".
  Otherwise, it is treated as a string.

Variables have a few benefits.
First, they let you move the assertions out from the normal program
code. In cases where you have many assertions on one line for example,
this can make the code easier to read & write.
Second, they enable more accurate comparisons in text probes.
Normally, when comparing a query and a string, the query is
"flattened" to a string. For node references, this means transforming
the reference to just the type of the node, which means you lose a
significant amount of information. When comparing two queries (which is
enabled by variables), the full query values are compared instead.

An example fictional source file with the new probes are shown below:

  void foo() { } // [[$foo:=Function]]

  void main() {
    foo();   // [[$ok:=Call]]
    bar();   // [[$bad:=Call]]
  }
  // [[$p:=Program]]

  // [[$ok.decl=$foo]]
  // [[$bad.decl=$p.unknownDecl]]

Variables in text probes are completely optional, you can continue using
simple string comparisons as before.

To prevent circular definition of variables, it is not permitted to
have a node reference on the right hand side of the = when defining a
new variable.