Shorten property names in UI
Added a setting to the CodeProber UI that automatically shortens
properties that do not match the regex /[\w\d\$_]+/, only including
the last part of the ID that does match the regex.
For example, "foo:bar:baz" becomes "baz".

When two names in a property list would shorten to the same label, the
full label is displayed instead.