AST view small improvements
Made the lines between AST nodes always get drawn. This fixes a problem
where if both AST nodes are out of bounds, but on opposite sides of the
window, then the lines would incorrectly disappear.

Updated the scrolling to be more natural. It should now scroll
more smoothly and in-line with other 3D software (more specifically I
compared with two different CAD-related programs). To mimic these other
pieces of software, the scroll direction is changed in this commit.