Refactor style.css
Extracted a number of custom variables in the css file, and made use of
them.
In theory this would make it easier to introduce new "themes" to
CodeProber in the future, instead of just light/dark. But this change is
mainly done to make the CSS file a bit nicer to work with.