Text probe bugfixes, improvements and tests
Fixed so nothing is highlighted when you hover "X" in [[A.b=X]] in a
textprobe. Since A.b is coerced to a String and compared with X, it
doesn't make sense to make X highlight things, even if it is a valid AST
type reference.

Made NodesWithProperty handle zero predicates (fixes a regression).

Made NodesWithProperty support the AST APIs cpr_cutoffNodeListTree.
This makes NodesWithProperty behave more like NodesAtPosition, and
improves the textprobe creation experience.

Wrote client-side tests for text probe hovering and completion.