Doc updates, windows bugfix
Moved the user-focused documentation in gradle_plugin/README to the
docs/ directory.

Wrote new documentation in gradle_plugin/README that focuses more on how
to develop the plugin itself.

Made carriage returns get removed at the end of strings in
EncodeResponseValue. It already removed newline characters, so this can
be seen as fixing a several year old Windows bug.