Aligned text probe handling
Made both the client and server implementation of text probes
behave the same for the following two cases:

1) It no longer accepts partial matches within '[[]]'.
   For example: "[[A ]]" would previously be accepted, but not anymore.

2) Text probes without assertions are now treated as implicit tests to
   check whether the query passes or not.

Both of these cases were previously handled differently by the client
and server, but they are now the same. Added regression tests to the
addnum workspace to make sure they stay aligned.

Made it possible to click a "failure" row in the "Run Tests" dialog to
go directly to the failing file.