Text probe array improvements
Made "~=" comparisons in text probe behave differently if the left side
is an array, and the right side is an AST node.
Previously, this resulted in substring comparisons.
Now, this results in exact node reference comparison.

This means that text probes on the following form may change behavior:
  [[A.b~=$c]]

This does *not* affect text probes on the following form:
  [[A.b~=C]]
This is because the right side of `=` is always treated as plain text
unless it starts with '$'.

To keep using the less precise contains-checking, you can add the name
of the right-side node rather than a variable reference to it. I.e.
instead of [[A.b~=$c]], do [[A.b~=MyNodeType]].

Additionally, text probe chaining will not automatically select nodes in
arrays if the array has exactly one member. For example, consider:
  [[A.b.c.d=E]]

If "c" returned an array, the text probe evaluation would previously
fail with an error about a broken reference/attribute chain.
Now, if the array has exactly one entry, it will be used.
If the array has multiple entries, the evaluation will fail with a
message about "multiple AST nodes".
The intention for this change is to make it easier to test for example
control-flow graph chains, where the common case is that each CFG node
has exactly one outgoing edge.

Added tests to the addnum/workspace that covers both these changes.